// src/app/interface/proposal.interface.ts
export interface ProposalList {
    policyNumber: string;
    policyStartDate: string;
    policyEndDate: string;
    intermediaryID: number;
    lastSyncTime: string;
    proposalNumber: number;
    firstName: string;
    updatedDate: string;
    lastName: string;
    policyType: string;
    policyStatus : string;
    productVarientName: string;
    mobileNo: number;
    totalPremiumInt: number;
    emailID: string;
  }